The 'Script' version of my program
This folder contains the version of the program that must be executed via a python distribution. Inside are all the auxillary files required for it to operate correctly and the older versions of the script, to show the production cycle of the software and the journey of its development, including the SQL scripts used to initially create and populate the database.