Create SB_setup.exe
This version of the program is wrapped up within an installer that will create both desktop and start menu shortcuts, as well as provide an uninstaller to make removal of the program quick and easy. The default install location is in Appdata/Roaming but users can specify their own install paths (IF YOU WISH TO INSTALL IT INTO PROGRAM FILES, YOU WILL NEED TO PROVIDE THE PROGRAM WITH ADMIN PRIVILEGES TO ENSURE IT SAVES CHANGES TO THE DATABASE AND THE LISTS THAT RUN THE SOFTWARE).